producer/requirements.txt
dnspython==2.8.0
pika==1.3.2
pymongo==4.15.3
python-dotenv==1.2.1
